Moved instance file location
