+++ instance/exampleConfig.js
