+++ src/index.js
@@ -1,6 +1,6 @@
 const { ModularClient } = require('./lib/client');
 
-const config = require('./instance/config');
+const config = require('../instance/config');
 
 const client = new ModularClient(config);
 client.login();
